Update thunderbird-stdlib (TB47 compatibility)
The previous version does not work with TB 47.
I have updated to the current version.
In this version msgHdrUtils.js needs misc.js, but this one needs a "../logs.js" which is not part of thunderbird-stdlib.
Therefore I have shortened the msgHdrUtils.js to only 1 function (the one I use)